Promote `statement` to be part of the regular CFG node API
... and use it from `TrackingPointMethods`
