+++ semanticcpg/src/main/scala/io/shiftleft/semanticcpg/language/nodemethods/TrackingPointMethods.scala
@@ -224,42 +224,17 @@ object TrackingPointMethodsBase {
 }
 
 object TrackingPointToCfgNode {
+  private val logger = LoggerFactory.getLogger(getClass)
   def apply(node: nodes.TrackingPointBase): nodes.CfgNode = {
     node match {
       case ftp: FakeTrackingPoint => ftp.trackedCfgNodeOverride
-      case _                      => applyInternal(node, _.parentExpression.get)
-    }
-
-  }
-
-  @scala.annotation.tailrec
-  private def applyInternal(node: nodes.TrackingPointBase,
-                            parentExpansion: nodes.Expression => nodes.Expression): nodes.CfgNode = {
-
-    node match {
-      case node: nodes.Identifier => parentExpansion(node)
-      case node: nodes.MethodRef  => parentExpansion(node)
-      case node: nodes.TypeRef    => parentExpansion(node)
-      case node: nodes.Literal    => parentExpansion(node)
-
-      case node: nodes.MethodParameterIn => node.method
-
-      case node: nodes.MethodParameterOut =>
-        node.method.methodReturn
-
-      case node: nodes.Call if MemberAccess.isGenericMemberAccessName(node.name) =>
-        parentExpansion(node)
-
-      case node: nodes.Call         => node
-      case node: nodes.ImplicitCall => node
-      case node: nodes.MethodReturn => node
-      case block: nodes.Block       =>
-        // Just taking the lastExpressionInBlock is not quite correct because a BLOCK could have
-        // different return expressions. So we would need to expand via CFG.
-        // But currently the frontends do not even put the BLOCK into the CFG so this is the best
-        // we can do.
-        applyInternal(TrackingPointMethodsBase.lastExpressionInBlock(block).get, identity)
-      case node: nodes.Expression => node
+      case cfgNode: nodes.CfgNode => cfgNode.statement
+      case _                      =>
+        // This case existed before, it just wasn't handled and would
+        // crash immediately instead.
+        logger.warn("Attempt to convert tracking point to CfgNode failed")
+        null
     }
   }
+
 }
